# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API estudiantes",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
        "/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Muestra un listado de estudiantes paginado.",
                "operationId": "App\\Http\\Controllers\\StudentController::index",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Código del estudiante",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Página de datos",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Cantidad de datos por página",
                        "schema": {
                            "type": "integer",
                            "default": 4
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Orden ascendiente (true) o descendiente (false)",
                        "schema": {
                            "type": "boolean",
                            "default": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un listado de estudiantes.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "per_page": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Student"
                                            },
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Permite crear un estudiante.",
                "operationId": "App\\Http\\Controllers\\StudentController::store",
                "requestBody": {
                    "description": "Objeto de solicitud de datos del estudiante.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2019-02-25"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "maxLength": 1,
                                        "example": "F"
                                    },
                                    "code": {
                                        "type": "string",
                                        "maxLength": 10,
                                        "example": "1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Devuelve una respuesta de creado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registro creado con éxito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/students/{id}": {
            "put": {
                "tags": [
                    "Students"
                ],
                "summary": "Permite crear un estudiante.",
                "operationId": "App\\Http\\Controllers\\StudentController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del estudiante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Objeto de solicitud de datos del estudiante.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2019-02-25"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "maxLength": 1,
                                        "example": "F"
                                    },
                                    "code": {
                                        "type": "string",
                                        "maxLength": 10,
                                        "example": "1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Devuelve una respuesta de actualizado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registro actualizado con éxito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "El registro solicitado no existe.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El registro solicitado no existe."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Student": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "string"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "string"
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2019-02-25"
                    },
                    "gender": {
                        "type": "string",
                        "maxLength": 1,
                        "example": "F"
                    },
                    "code": {
                        "type": "string",
                        "maxLength": 10,
                        "example": "1234567890"
                    },
                    "full_name": {
                        "type": "string",
                        "example": "string"
                    },
                    "age": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Student"
                }
            }
        }
    },
    "security": [
        []
    ]
}